refactor: move functions to utils,use iife
